fix validation
