+++ models/contacts.json
@@ -58,5 +58,17 @@
     "name": "Alec Howardy",
     "email": "howard@gmail.com",
     "phone": "(555) 999-9999"
+  },
+  {
+    "id": "QodD57zXv0-uON8NqpvAw",
+    "name": "Vladir Zelensky",
+    "email": "zelensky@gmail.com",
+    "phone": "(079)-999-9999"
+  },
+  {
+    "id": "tbQfbh9l3Rt14BjOIlFqu",
+    "name": "Vladir Zelensky",
+    "email": "zelensky@gmail.com",
+    "phone": "(333)-999-9999"
   }
 ]
